Fix units init-company and init-company-statistics
Fix company-force-prefix.
Because company-capf may returns candidates whose prefix is slightly
different from input prefix, this transformer was added to force
company candidates have the input one.
Because other backend, e.g., company-files, may not work with this
transformer, it was fixed as it performs filtering only when using
company-capf.